+++ lua/typewriter/logger.lua
@@ -5,6 +5,25 @@
 
 local log_file = vim.fn.stdpath('data') .. '/typewriter.log'
 
+local M = {}
+
+--- Set a custom log file path
+--
+-- Allows tests or users to override the default location.
+-- The directory of the provided path will be created automatically
+-- when writing messages.
+-- @param path string new log file path
+function M.set_log_file(path)
+  log_file = path
+end
+
+--- Get the current log file path
+-- @return string
+function M.get_log_file()
+  return log_file
+end
+
+
 --- Append a log message to the log file, creating the directory if needed.
 -- @param level string log level label
 -- @param msg string message to write
@@ -20,8 +39,6 @@ local function write(level, msg)
     file:close()
   end
 end
-
-local M = {}
 
 function M.info(msg)
   write('INFO', msg)