+++ spec/spec_helper.lua
@@ -7,10 +7,13 @@ vim.fn = vim.fn or {}
 vim.api.nvim_exec_autocmds = vim.api.nvim_exec_autocmds or function() end
 vim.api.nvim_win_set_cursor = vim.api.nvim_win_set_cursor or function() end
 vim.api.nvim_win_get_cursor = vim.api.nvim_win_get_cursor or function() return {1,0} end
+vim.api.nvim_create_autocmd = vim.api.nvim_create_autocmd or function() end
+vim.api.nvim_create_user_command = vim.api.nvim_create_user_command or function() end
 vim.cmd = vim.cmd or function() end
 vim.schedule = vim.schedule or function(fn) fn() end
 vim.log = vim.log or { levels = { INFO = 1 } }
 vim.notify = vim.notify or function() end
+vim.v = vim.v or {}
 vim.fn.stdpath = vim.fn.stdpath or function() return '/tmp' end
 vim.fn.fnamemodify = vim.fn.fnamemodify or function(path, mod)
   if mod == ':h' then